TestHelpers API

Funciones para pruebas del paquete. Los datos de Macronia de esta documentación son generados con estas utilidades.

CPIDataBase.TestHelpers.getrandomweights — Function

getrandomweights(T=Float32, G=218)

Función para generar pesos aleatorios

CPIDataBase.TestHelpers.getbasedates — Function

getbasedates(vmat, startdate=Date(2000, 12))

Función para generar fechas a partir de matriz de variaciones intermensuales

CPIDataBase.TestHelpers.getzerobase — Function

getzerobase(; 
    [T_type = Float32, 
    G = 218, 
    T_periods = 120, 
    startdate = Date(2001,1), 
    baseindex = 100*one(T_type)])

Función para obtener base de tipo VarCPIBase con variaciones intermensuales iguales a cero.

CPIDataBase.TestHelpers.getrandombase — Function

getrandombase(; 
    [T_type = Float32, 
    G = 218, 
    T_periods = 120, 
    startdate = Date(2001,1), 
    baseindex = 100*one(T_type)])

Función para obtener una base de tipo VarCPIBase con variaciones intermensuales aleatorias.

CPIDataBase.TestHelpers.getzerocountryst — Function

getzerocountryst(T_type = Float32)

Obtiene un UniformCountryStructure cuyas variaciones intermensuales son todas iguales a cero en la configuración de períodos del IPC de Guatemala.

CPIDataBase.TestHelpers.getrandomcountryst — Function

getrandomcountryst(T_type = Float32)

Obtiene un UniformCountryStructure cuyas variaciones intermensuales son todas aleatorias en la configuración de períodos del IPC de Guatemala.